app/requirements.txt
fastapi
tensorflow
transformers
uvicorn
pydantic
tf-keras